+++ iop/fs/bdm/Makefile
@@ -8,7 +8,7 @@
 
 # IOP_CFLAGS += -DDEBUG
 
-IOP_OBJS = main.o bdm.o part_driver.o imports.o exports.o
+IOP_OBJS = main.o bdm.o part_driver.o part_driver_mbr.o part_driver_gpt.o imports.o exports.o
 IOP_LIB_ARCHIVES = $(PS2SDKSRC)/iop/fs/libbdm/lib/libbdm.a
 IOP_CFLAGS = -I$(PS2SDKSRC)/iop/fs/libbdm/include/ -Wno-strict-aliasing
